update the display of the total items on the home page in a js separate file
