+++ src/modules/totalItems.js
@@ -1,0 +1,12 @@
+const totalBooksItems = async () => {
+  const options = ['html', 'css', 'javascript', 'ruby', 'react', 'node', 'jokes', 'java', 'maths', 'art', 'spanish', 'english', 'python', 'sql'];
+  const num = Math.floor(Math.random() * 13);
+  const search = document.getElementById('search').value || options[num];
+  const url = `https://www.googleapis.com/books/v1/volumes?q=${search}`;
+  const res = await fetch(url);
+  const data = await res.json();
+  const { totalItems } = data;
+  return { data, totalItems, search };
+};
+
+export default totalBooksItems;
