+++ src/style.css
@@ -4,10 +4,13 @@ body {
 }
 
 header {
-  height: 100px;
   display: flex;
   justify-content: space-between;
   border-bottom: 2px solid gray;
+}
+
+nav {
+  width: 100%;
 }
 
 .header-img {
